requirements.txt
# =========================================
# 🧠 STOCK SCANNER – Gesamt-Requirements
# =========================================

# Basis / Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0
joblib>=1.3.0
pyarrow>=17.0.0     # Parquet-Support (für Caching)
lxml>=4.9.0
html5lib>=1.1
feedparser>=6.0.11
matplotlib>=3.8.0   # Für Backtests, optional Plot-Ausgaben

# Finanzdaten
yfinance>=0.2.40

# Machine Learning / KI
scikit-learn>=1.5.0
xgboost>=2.1.0
lightgbm>=4.3.0
tensorflow>=2.16.1  # stabiler als 2.20.0 (kompatibler mit scikit-learn)
torch>=2.3.0
transformers>=4.41.0
sentence-transformers>=3.0.0
optuna>=3.5.0

# Sentiment & NLP
nltk>=3.8.1
vaderSentiment>=3.3.2

# OpenAI / API
openai>=1.3.0

# Optional / Nützlich
beautifulsoup4>=4.12.0
